refactor: util 추가 (#245)
DTO에서 피드 상태값을 조회해야해서 Util로 해당 기능을 이동시킴
